Fix/scroll

+++ include/lazarus_window_manager.h
@@ -83,7 +83,7 @@ class Events
 		uint32_t mousePositionX;
 		uint32_t mousePositionY;
 		
-		int8_t scrollState;
+		float scrollState;
 
         virtual ~Events();
